add ads image upload

+++ src/hooks/useAds.ts
@@ -1,6 +1,5 @@
 'use client';
 import { useRouter } from "next/navigation";
-// import { uploadGoogleFile } from "@/src/hooks/google_settings";
 
 export default function useAds() {
     const router = useRouter();
@@ -17,8 +16,6 @@ export default function useAds() {
             // upload image to cloud and get the url
             const formData = new FormData();
             formData.append("file", picture);
-            console.log("formData: ", formData);
-            console.log("picture: ", picture);
 
             const imageRes = await fetch("/api/image", {
                 method: "POST",
@@ -29,28 +26,25 @@ export default function useAds() {
                 alert("Error uploading image");
                 throw new Error(error);
             }
-
-            // const imageUrl = await imageRes.json();
-
-            // console.log("imageUrl: ", imageUrl);
+            const imageUrl = await imageRes.json();
 
             // upload the image url and language to the database
-            // const res = await fetch("/api/ads", {
-            //     method: "POST",
-            //     headers: {
-            //         "Content-Type": "application/json",
-            //     },
-            //     body: JSON.stringify({
-            //         picture: imageUrl, 
-            //         language: language,
-            //     }),
-            // });
-            // if (!res.ok) {
-            //     const body = await res.json();
-            //     throw new Error(body.error);
-            // }
-            // 
-            // router.refresh();
+            const res = await fetch("/api/ads", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify({
+                    picture: imageUrl.url, 
+                    language: language,
+                }),
+            });
+            if (!res.ok) {
+                const body = await res.json();
+                throw new Error(body.error);
+            }
+            
+            router.refresh();
         
         } catch (error) {
             console.error("error: ", error);
@@ -102,6 +96,38 @@ export default function useAds() {
 
     //DELETE
     const deleteAds = async (id: number) => {
+        // get the image url from the database
+        const resGet = await fetch("/api/ads", {
+            method: "GET",
+            headers: {
+                "Content-Type": "application/json",
+            },
+        });
+        if (!resGet.ok) {
+            const body = await resGet.json();
+            throw new Error(body.error);
+        }
+        const adsList = await resGet.json();
+        const ads = adsList.ads;
+        const target = ads.find((ad: any) => ad.id === id);
+        const url = target.picture;
+
+        // delete image from cloud
+        const resDelete = await fetch("/api/image", {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+                url,
+            }),
+        });
+        if (!resDelete.ok) {
+            const body = await resDelete.json();
+            throw new Error(body.error);
+        }
+
+        // delete the object from the database
         const res = await fetch('/api/ads', {
             method: "DELETE",
             headers: {
